Add generic selector
Remove duplication by making more generic selector that includes any col- class in any part of the class list in the class attribute. This is based off of #40fdee1 commit that Matt Roach started that Jack Howard reverted after it did not apply to tags that did not begin with col-, like the footer.